more awesomeness :dancer:
This commit adds a event sourced recipe service. It turns commands into
an event log and stores it into a PostgreSQL DB. Notifications about
updated recipes a distributed through Kafka. Views are built right away
and stored into ElasticSearch

ElasticSearch gives us amazing search capabilities. The reason why we
are using both Postgres and Elastic is because they are good at diffrent
things. Elastic is fast but not so durable, Postgres is super durable
but hard to search within.

Also more UI updates. New discover page and Find page. While the
discover page is fully mocked as of now, the find page is actually
wokring for real :star:

Le chef! We have a new companion in our team by the name Le Chef. He is
a master at comming up with intresting recipes in a very short period of
time.

We also have auth in the frontend, but we still need to enforce it in
the backend. Most resolver are open, but mutations are often gated
behind login. Examples are createRecipe setDescription etc.

Hope you enjoy!